i messed up git, now i have a headache

+++ Core/Src/rf.h
@@ -2,6 +2,6 @@
 #include <stdio.h>
 #include <stdint.h>
 
-void rf_send_command(uint16_t cmd);
+void rf_send_command(uint8_t cmd);
 void rf_setup_commands();
 void init_rf();
